feat(update all request): expand config options and abstract to extensions
