+++ Turner.Infrastructure.Crud/Configuration/Builders/CrudRequestEntityConfigBuilder.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using System.Linq.Expressions;
 using System.Threading.Tasks;
 using Turner.Infrastructure.Crud.Configuration.Builders.Filter;
 using Turner.Infrastructure.Crud.Configuration.Builders.Select;
@@ -224,80 +222,7 @@ namespace Turner.Infrastructure.Crud.Configuration.Builders
 
             return this;
         }
-
-        public CrudRequestEntityConfigBuilder<TRequest, TEntity> UpdateAllWith<TIn>(
-            Expression<Func<TRequest, IEnumerable<TIn>>> requestEnumerableExpr,
-            string requestItemKeyProperty,
-            string entityKeyProperty,
-            Func<TIn, TEntity, TEntity> updator)
-        {
-            FilterWith(builder =>
-                builder.FilterOnCollection(requestEnumerableExpr, requestItemKeyProperty, entityKeyProperty));
-
-            // TODO: Move this to a builder
-            // TODO: Default to Mapper.Map for updator
-            // TODO: Support Expression key properties
-            // TODO: Support the key comparer version of Join?
-            // TODO: Debugging features (active filters, etc)
-
-            var rParamExpr = Expression.Parameter(typeof(TRequest));
-            
-            var eParamExpr = Expression.Parameter(typeof(TEntity));
-            var eKeyExpr = Expression.PropertyOrField(eParamExpr, entityKeyProperty);
-
-            var enumerableMethods = typeof(Enumerable).GetMethods();
-
-            var tupleType = typeof(Tuple<,>).MakeGenericType(typeof(TIn), typeof(TEntity));
-            var tupleCtor = tupleType.GetConstructor(new[] { typeof(TIn), typeof(TEntity) });
-            
-            var selectTupleParam = Expression.Parameter(tupleType);
-            var item1Param = Expression.PropertyOrField(selectTupleParam, "Item1");
-            var item2Param = Expression.PropertyOrField(selectTupleParam, "Item2");
-            var updateExpr = Expression.Call(updator.Method, item1Param, item2Param);
-            var selectLambdaExpr = Expression.Lambda(updateExpr, selectTupleParam);
-            
-            var esParamExpr = Expression.Parameter(typeof(TEntity[]));
-            
-            var joinInfo = enumerableMethods
-                .Single(x => x.Name == "Join" && x.GetParameters().Length == 5)
-                .MakeGenericMethod(typeof(TEntity), typeof(TIn), eKeyExpr.Type, tupleType);
-            var joinEntityParam = Expression.Parameter(typeof(TEntity));
-            var joinEntityKeyExpr = Expression.Lambda(
-                Expression.PropertyOrField(joinEntityParam, entityKeyProperty),
-                joinEntityParam);
-            var joinInParam = Expression.Parameter(typeof(TIn));
-            var joinInKeyExpr = Expression.Lambda(
-                Expression.PropertyOrField(joinInParam, requestItemKeyProperty),
-                joinInParam);
-            var joinOutEntityParam = Expression.Parameter(typeof(TEntity));
-            var joinOutInParam = Expression.Parameter(typeof(TIn));
-            var joinOutExpr = Expression.Lambda(
-                Expression.New(tupleCtor, joinOutInParam, joinOutEntityParam), 
-                joinOutEntityParam, 
-                joinOutInParam);
-            var joinInEnumParam = Expression.Invoke(requestEnumerableExpr, rParamExpr);
-            var joinExpr = Expression.Call(joinInfo, esParamExpr, joinInEnumParam, joinEntityKeyExpr, joinInKeyExpr, joinOutExpr);
-
-            var selectInfo = enumerableMethods
-                .Single(x => x.Name == "Select" &&
-                                x.GetParameters().Length == 2 &&
-                                x.GetParameters()[1].ParameterType.GetGenericArguments().Length == 2)
-                .MakeGenericMethod(tupleType, typeof(TEntity));
-            var selectExpr = Expression.Call(selectInfo,  joinExpr, selectLambdaExpr);
-
-            var toArrayInfo = enumerableMethods
-                .Single(x => x.Name == "ToArray" && x.GetParameters().Length == 1)
-                .MakeGenericMethod(typeof(TEntity));
-            var toArrayExpr = Expression.Call(toArrayInfo, selectExpr);
-
-            var lambdaExpr = Expression.Lambda(toArrayExpr, rParamExpr, esParamExpr);
-            var lambda = (Func<TRequest, TEntity[], TEntity[]>) lambdaExpr.Compile();
-
-            _updateEntitiesFromRequest = (request, entities) => Task.FromResult(lambda(request, entities));
-
-            return this;
-        }
-
+        
         public void Build<TCompatibleRequest>(CrudRequestConfig<TCompatibleRequest> config)
         {
             if (_optionsConfig != null)
